Updated mapping to correct location after folder restructuring
