+++ client/src/components/PhotoItem/PhotoItem.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 const PhotoItem = (props) => {
   const { photo } = props;
-  const { thumbnailUrl, imageUrl, description } = photo;
+  const { id, thumbnailUrl, imageUrl, description } = photo;
 
   return (
-    <img src={thumbnailUrl} alt={description} />
+    <img src={thumbnailUrl} alt={description} photo-id={id} />
   );
 };
